Add scene and world modules; implement LightList and random scene generation
