+++ src/material.rs
@@ -2,7 +2,7 @@ use crate::color::Color;
 use crate::hittable::HitRecord;
 use crate::light::Light;
 use crate::ray::Ray;
-use crate::vec3::{Point3, Vec3};
+use crate::vec3::Point3;
 use crate::{common, vec3};
 
 pub trait Material: Send + Sync {
